+++ mainS.py
@@ -1,7 +1,10 @@
 # -*- coding:utf-8 -*-
 from src.Server.MessageTransferSystem.VersionControlSystem import versionControl
 
+import os
+
 # 主界面
 if __name__ == '__main__':
     versionControl.VersionControl()
+
