+++ mainU.py
@@ -1,7 +1,11 @@
 # -*- coding:utf-8 -*-
-from src.Client.VersionControlSystem.GetNewFile import doUpdate
+import os
+from tkinter import messagebox
+from src.Update.VersionControlSystem import versionControl
 
 # 主界面
 if __name__ == '__main__':
-    doUpdate.GetNewFile()
+    versionControl.VersionControl().handle()
+    command = 'MemoryAssist v1.0.exe'
+    os.system(command)
