+++ src/Update/VersionControlSystem/GetNewFile/Tools/getIndexFile.py
@@ -1,5 +1,5 @@
 # -*- coding:utf-8 -*-
-from src.Client.Conf.config import *
+from src.Update.Conf.config import *
 
 
 class GetIndexFile():
