+++ src/Update/VersionControlSystem/JudgeNeedUpdate/judgeNeedUpdate.py
@@ -1,5 +1,5 @@
 # -*- coding:utf-8 -*-
-from src.Client.Conf.config import *
+from src.Update.Conf.config import *
 from src.Client.SystemTools.ConfFileRead import configFileRead
